requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dataforseo-client==2.0.10
deprecation==2.1.0
filelock==3.19.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==2.21.1
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.21.1
requests==2.32.5
requests-file==2.1.0
six==1.17.0
sniffio==1.3.1
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.2
websockets==15.0.1
